Clarifies the usage of mrobosub_lib.Node
Most of the doc comments provided are not coherent with how we are using
Node currently. I agree that we should keep mrobosub_lib.Node as a
common ancestor of all nodes that we have. That said the comments were
regarding using Node in a way that we (verbally) decided against after
the comments were made. So I clarified them to be coherent with how we
are using mrobosub_lib.Node right now. Later on, if this changes, doc
comments can be added again.